Demos/MQTT FIS Simulator/requirements.txt
paho-mqtt==2.0.0
requests==2.31.0

